Frida Kahlo Retrospective Off-Platform Project
Overview

This project is slightly different than others you have encountered thus far on Codecademy. Instead of a step-by-step tutorial, this project contains a series of open-ended requirements which describe the project you’ll be building.

Project Goals

Congratulations! You’ve been hired to work on a retrospective of Frida Kahlo’s work at a major museum in Mexico.
Your job is to put together the audio tour, but in order to do that, you need to create a list of each painting featured in the exhibit, the date it was painted, and its spot on the tour.
Use your knowledge of Python lists to create a master list of each painting, its date, and its audio tour ID.